+++ changelog.md
@@ -1,4 +1,8 @@
 # Moni Labs 0.10.0
 
-* Update to gtm 7.1.2
+* Update to gtm 7.1.3
 * Add a bunch of placeholders for the central monitor system
+* Fix a variety of bugs
+* Add EMI info to microverse projector recipes
+* Make antimatter collider recipes no longer hardcoded
+* Slow down Sculk Vat xp buffer decay